docs: add clarification to pagination.mdx (#336)
As I was using elf state management in my project I referenced the docs, and because of line 55 in pagination.mdx I assumed that elf pagination started at index 1. This cause a lot of extra work for myself, I did not even try to check if there was a page "0". I just wanted to propose some sort of clarification to the docs so that first time users of elf are aware that they can start their pagination at index 0.