feat(state-history): 🔥 Add an option to control future state
An option called resetFutureOnNewState has been added to the
StateHistoryOptions interface. This option allows the user to control
whether the future redo states should be cleared when a new state is
added after the user has undone one or more state changes. The default
value of this option is set to false, meaning that by default, the
future states will not be cleared.